Introduce BatchSpanExportProcessorOptions
