+++ src/OpenTelemetry.Exporter.Jaeger/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Unreleased
 
+* `JaegerExporterOptions.BatchExportProcessorOptions` is initalized with
+  `BatchSpanExportProcessorOptions` which supports field value overriding
+  using `OTEL_BSP_SCHEDULE_DELAY`, `OTEL_BSP_EXPORT_TIMEOUT`,
+  `OTEL_BSP_MAX_QUEUE_SIZE`, `OTEL_BSP_MAX_EXPORT_BATCH_SIZE`
+  envionmental variables as defined in the
+  [specification](https://github.com/open-telemetry/opentelemetry-specification/blob/v1.5.0/specification/sdk-environment-variables.md#batch-span-processor).
+  ([#2219](https://github.com/open-telemetry/opentelemetry-dotnet/pull/2219))
+
 ## 1.2.0-alpha1
 
 Released 2021-Jul-23
