+++ src/OpenTelemetry.Exporter.Jaeger/JaegerExporterOptions.cs
@@ -18,6 +18,7 @@ using System;
 using System.Diagnostics;
 using System.Security;
 using OpenTelemetry.Exporter.Jaeger.Implementation;
+using OpenTelemetry.Trace;
 
 namespace OpenTelemetry.Exporter
 {
@@ -82,6 +83,6 @@ namespace OpenTelemetry.Exporter
         /// <summary>
         /// Gets or sets the BatchExportProcessor options. Ignored unless ExportProcessorType is BatchExporter.
         /// </summary>
-        public BatchExportProcessorOptions<Activity> BatchExportProcessorOptions { get; set; } = new BatchExportProcessorOptions<Activity>();
+        public BatchExportProcessorOptions<Activity> BatchExportProcessorOptions { get; set; } = new BatchSpanExportProcessorOptions();
     }
 }
